% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmaker.R
\name{custom_plot}
\alias{custom_plot}
\title{Creates a scatter plot given the data column in field data.}
\usage{
custom_plot(data, column_one, column_two, column_three, plot_type)
}
\arguments{
\item{column_one}{the name of the column for the independent variable}

\item{column_two}{the name of the column for the dependent variable}

\item{column_three}{the name of the column for facet plot sections}

\item{data_name}{the name of the data frame you wish to use}
}
\value{
a scatter plot that shows independent and dependent variables
}
\description{
Utilizes the ggplot2 package and the function geom_scatter to plot field data.
Returns a plot that illustrates the relationship of the variables/data.
}
