% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamean.R
\name{data_mean}
\alias{data_mean}
\title{calculates mean of more than one column and places into new column}
\usage{
data_mean(data_name, column)
}
\arguments{
\item{data_name}{Name of the data set you are working within}

\item{column}{Name of the column(s) you wish to calculate the mean for}

\item{new_column}{Name of a new column to place calculations in}
}
\value{
calculations of mean for the specified column (numeric)
}
\description{
Uses the mean function of the specified column to calculate the mean of the column and places new mean calculation into a new column.
Returns data set with new column filled with the calculated mean of a column.
}
