% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amphibiandatalm.R
\name{lm_model}
\alias{lm_model}
\title{Creates a linear model using two predictor variables}
\usage{
lm_model(data_name, column_one, column_two, column_three)
}
\arguments{
\item{data_name}{the name of the data set being used}

\item{column_one}{the response variable desired}

\item{column_two}{the first predictor variable}

\item{column_three}{the second predictor variable}
}
\value{
a linear model summary
}
\description{
Uses the lm function to analyze and create linear models for data sets with more than one predictor variable of a desired response variable.
Returns a linear model summary that illustrates significance.
}
