% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manovamaker.R
\name{manova_maker}
\alias{manova_maker}
\title{Creates a multivariate analysis on field data}
\usage{
manova_maker(data_name, column_one, column_two, column_three)
}
\arguments{
\item{data_name}{name of the data set being used}

\item{column_one}{the name of the column being used as the response variable}

\item{column_two}{the name of the column being used as the second response variable}

\item{column_three}{the name of the column being used as the predictor variable}
}
\value{
a density plot for the columns specified
}
\description{
Utilizes the manova function to make a multivariate analysis.
Return a summary fit indicating the variance of the data set among variables.
}
