% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapcolumns.R
\name{scrap_col}
\alias{scrap_col}
\title{Scraps the unnecessary columns in the data that is not needed for manipulations}
\usage{
scrap_col(data_name, column_list)
}
\arguments{
\item{data_name}{name of the data set used}

\item{column_list}{name of column(s) that aren't useful}
}
\value{
data set without the specified columns
}
\description{
Uses select function to get rid of columns that are not useful in the analyses
Returns a data set without the columns inputted as an argument
}
